+++ common/linux_perf_hpm.c
@@ -53,52 +53,52 @@ int linuxPerfCntSet(){
 }
 
 int linuxPerfCntGet(){
-  size_t instret, cycles;
+  long instret, cycles;
 
   instret = read_csr(minstret);
   cycles = read_csr(mcycle);
 
-  printf("instructions:%d\n", (int)(instret));
-  printf("cpu_cycle:%d\n", (int)(cycles));
+  printf("instructions:%ld\n", (long)(instret));
+  printf("cpu_cycle:%ld\n",    (long)(cycles));
 
-  SHOW_PERFCNT("l1_icache_loads:%d\n", 3);
-  SHOW_PERFCNT("l1_icache_load_misses:%d\n", 4);
-  SHOW_PERFCNT("l1_icache_prefetches:%d\n", 5);
-  SHOW_PERFCNT("l1_icache_prefetch_misses:%d\n", 6);
+  SHOW_PERFCNT("l1_icache_loads:%ld\n", 3);
+  SHOW_PERFCNT("l1_icache_load_misses:%ld\n", 4);
+  SHOW_PERFCNT("l1_icache_prefetches:%ld\n", 5);
+  SHOW_PERFCNT("l1_icache_prefetch_misses:%ld\n", 6);
 
-  SHOW_PERFCNT("l1_dcache_loads:%d\n", 7);
-  SHOW_PERFCNT("l1_dcache_load_misses:%d\n", 8);
-  SHOW_PERFCNT("l1_dcache_stores:%d\n", 9);
-  SHOW_PERFCNT("l1_dcache_store_misses:%d\n", 10);
-  SHOW_PERFCNT("l1_dcache_prefetches:%d\n", 11);
-  SHOW_PERFCNT("l1_dcache_prefetch_misses:%d\n", 12);
+  SHOW_PERFCNT("l1_dcache_loads:%ld\n", 7);
+  SHOW_PERFCNT("l1_dcache_load_misses:%ld\n", 8);
+  SHOW_PERFCNT("l1_dcache_stores:%ld\n", 9);
+  SHOW_PERFCNT("l1_dcache_store_misses:%ld\n", 10);
+  SHOW_PERFCNT("l1_dcache_prefetches:%ld\n", 11);
+  SHOW_PERFCNT("l1_dcache_prefetch_misses:%ld\n", 12);
 
-  SHOW_PERFCNT("itlb_loads:%d\n", 13);
-  SHOW_PERFCNT("itlb_load_misses:%d\n", 14);
-  SHOW_PERFCNT("itlb_prefetches:%d\n", 15);
-  SHOW_PERFCNT("itlb_prefetch_misses:%d\n", 16);
+  SHOW_PERFCNT("itlb_loads:%ld\n", 13);
+  SHOW_PERFCNT("itlb_load_misses:%ld\n", 14);
+  SHOW_PERFCNT("itlb_prefetches:%ld\n", 15);
+  SHOW_PERFCNT("itlb_prefetch_misses:%ld\n", 16);
 
-  SHOW_PERFCNT("dtlb_loads:%d\n", 17);
-  SHOW_PERFCNT("dtlb_load_misses:%d\n", 18);
-  SHOW_PERFCNT("dtlb_stores:%d\n", 19);
-  SHOW_PERFCNT("dtlb_store_misses:%d\n", 20);
-  SHOW_PERFCNT("dtlb_prefetches:%d\n", 21);
-  SHOW_PERFCNT("dtlb_prefetch_misses:%d\n", 22);
+  SHOW_PERFCNT("dtlb_loads:%ld\n", 17);
+  SHOW_PERFCNT("dtlb_load_misses:%ld\n", 18);
+  SHOW_PERFCNT("dtlb_stores:%ld\n", 19);
+  SHOW_PERFCNT("dtlb_store_misses:%ld\n", 20);
+  SHOW_PERFCNT("dtlb_prefetches:%ld\n", 21);
+  SHOW_PERFCNT("dtlb_prefetch_misses:%ld\n", 22);
 
-  SHOW_PERFCNT("cache_references:%d\n", 23); // llc references -> hit
-  SHOW_PERFCNT("cache_misses:%d\n", 24);     // llc misses
-  SHOW_PERFCNT("llc_cache_loads:%d\n", 25);
-  SHOW_PERFCNT("llc_cache_load_misses:%d\n", 26);
-  SHOW_PERFCNT("llc_cache_stores:%d\n", 27);
-  SHOW_PERFCNT("llc_cache_store_misses:%d\n", 28);
-  SHOW_PERFCNT("llc_cache_prefetches:%d\n", 29);
-  SHOW_PERFCNT("llc_cache_prefetch_misses:%d\n", 30);
+  SHOW_PERFCNT("cache_references:%ld\n", 23); // llc references -> hit
+  SHOW_PERFCNT("cache_misses:%ld\n", 24);     // llc misses
+  SHOW_PERFCNT("llc_cache_loads:%ld\n", 25);
+  SHOW_PERFCNT("llc_cache_load_misses:%ld\n", 26);
+  SHOW_PERFCNT("llc_cache_stores:%ld\n", 27);
+  SHOW_PERFCNT("llc_cache_store_misses:%ld\n", 28);
+  SHOW_PERFCNT("llc_cache_prefetches:%ld\n", 29);
+  SHOW_PERFCNT("llc_cache_prefetch_misses:%ld\n", 30);
 
-  // SHOW_PERFCNT("bus_cycles:%d\n", 31);
-  // SHOW_PERFCNT("branches:%d\n", 29);
-  // SHOW_PERFCNT("branch_misses:%d\n", 30);
-  // SHOW_PERFCNT("branch_loads:%d\n", 31);
-  // SHOW_PERFCNT("branch_load_misses:%d\n", 32);
+  // SHOW_PERFCNT("bus_cycles:%ld\n", 31);
+  // SHOW_PERFCNT("branches:%ld\n", 29);
+  // SHOW_PERFCNT("branch_misses:%ld\n", 30);
+  // SHOW_PERFCNT("branch_loads:%ld\n", 31);
+  // SHOW_PERFCNT("branch_load_misses:%ld\n", 32);
 
   return 0;
 }
